Move VertexAIError implementation away from types
